feat: dynamic golangci-lint versions and configuration (#145)
https://typeform.atlassian.net/browse/PLT-2300

Make the configuration file location of golangci-lint (golangci.yaml) configurable.

This is a preparation for the move to golangci-lint v2. The configuration files between v1 and v2 are not compatible, which means that once we move the golangci-lint executable to 2.x, a single location for the file is not sufficient, as some repos will be on v1, and others on v2.

The defaults are the hardcoded values from today, with the exception of ref, which is set to 1.15.0 (latest tag with a v1 configuration file).

Tested for backwards compatibility: https://github.com/Typeform/event-dispatcher/actions/runs/17552732900/job/49848910468?pr=605