fix(deep-purple): more reliable check for PR comments (#158)
https://typeform.atlassian.net/browse/PLT-2487

- Add the `deep-purple-checks` PR comments only on `pull_request` and
`pull_request_target` events
- Add the `deep-purple-checks` PR comments if the workflow has not been
canceled, as recommended [in Github
docs](https://docs.github.com/en/actions/reference/workflows-and-actions/expressions#always).

The reason for this change is that this positive check is more reliable,
and supports excluding runs on a merge group (where no comments can be
added), for example.

`pull_request_target` is only used by
https://github.com/Typeform/followups-manager/blob/6b9546d94fb4d4bb5b5f0efc45afc52d3885d87e/.github/workflows/dependabot.yml#L4
which just auto-approves dependabot PRs, and therefore wouldn't ever run
the deep-purple-checks. However, as the event is published in a PR, a
comment can be added, so we can easily support it here as well. If it's
ever used, the problem won't be the PR comments, but Deep Purple runs
themselves.